refactor(resilience): consolidate circuit breakers; deprecate v2 via shim (5.2)
Three circuit breaker implementations coexisted in aragora/resilience/.
This consolidates to two canonical homes and makes circuit_breaker_v2 a
deprecation shim, with zero behavior change (preservation rule: no deletion,
implementation relocated verbatim):

- simple_circuit_breaker.py is now the canonical home of the thread-safe
  sync-core breakers: SimpleCircuitBreaker plus the configurable
  BaseCircuitBreaker (success thresholds, half-open gating, excluded
  exceptions, failure-rate windows, async/sync decorators, named registry)
  moved verbatim from circuit_breaker_v2.py.
- circuit_breaker.py remains the asyncio-native multi-entity CircuitBreaker.
- circuit_breaker_v2.py is now an explicit re-export shim that emits a
  DeprecationWarning on import; all public names and registry internals are
  identity-equal to their canonical targets (shared state, isinstance-safe).

Why relocation instead of re-mapping onto CircuitBreaker/SimpleCircuitBreaker:
BaseCircuitBreaker has distinct half-open semantics (consecutive
success_threshold + half_open_max_requests gating, excluded_exceptions,
failure-rate window, CircuitBreakerStats) that neither target replicates;
both production call-sites depend on them, so re-mapping would have silently
changed failure behavior.

Importers migrated (semantics exact, same objects):
- aragora/resilience/__init__.py (package re-exports, no longer triggers
  the shim warning on package import)
- aragora/agents/external/base.py (sync surface)
- aragora/server/handlers/composite.py (sync registry read)

Tests: new tests/resilience/test_cb_v2_deprecation.py (17 tests: warning
fires once, re-exports identity-equal, registry shared, package import does
not warn). Repointed implementation-internal patch targets
(circuit_breaker_v2.time, sys.modules import-failure simulations) to the new
home; left tests/resilience_patterns and
tests/test_resilience_patterns_circuit_breaker importing via the shim as
live compat exercisers. tests/resilience 826 passed, subsystem slices 514
passed, api_agents 784 passed; ruff + mypy clean.
